backend/requirements.txt
google-generativeai
python-dotenv
Flask
Flask-Cors
